Remove &nbsp; hack from page nav
An empty div normally won't take up space, which was a problem on
the first and last chapters as the chapter links got messed up. My
previous solution inserted a &nbsp; into the div, which worked, but
also caused there to be a weird selectable space in the div.

Setting min-height achieves the same effect without any selectable
text.